pyproject.toml
[project]
name = "Orunmila"
version = "0.1.0"
description = "An AI agent that answers questions about Yoruba history and culture, integrated with Telex.im"
requires-python = ">=3.11.9"
dependencies = [
    "fastapi[all]",
    "pydantic-ai",
    "pydantic-settings",
    "httpx",
    "python-dotenv",
    "uvicorn[standard]",
    "redis[hiredis]",
    "jsonrpcclient",
    "cairosvg",
    "minio"
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
